refactor(commands): move shared helpers into modules
